+++ packages/orq-rc/src/orq_ai_sdk/models/createpromptop.py
@@ -30,9 +30,6 @@ r"""The modality of the model"""
 CreatePromptFormat = Literal["url", "b64_json", "text", "json_object"]
 r"""Only supported on `image` models."""
 
-Quality = Literal["standard", "hd"]
-r"""Only supported on `image` models."""
-
 CreatePromptResponseFormatType = Literal["json_object"]
 
 
@@ -131,7 +128,7 @@ class ModelParametersTypedDict(TypedDict):
     r"""Only supported on `image` models."""
     dimensions: NotRequired[str]
     r"""Only supported on `image` models."""
-    quality: NotRequired[Quality]
+    quality: NotRequired[str]
     r"""Only supported on `image` models."""
     style: NotRequired[str]
     r"""Only supported on `image` models."""
@@ -193,7 +190,7 @@ class ModelParameters(BaseModel):
     dimensions: Optional[str] = None
     r"""Only supported on `image` models."""
 
-    quality: Optional[Quality] = None
+    quality: Optional[str] = None
     r"""Only supported on `image` models."""
 
     style: Optional[str] = None
@@ -623,9 +620,6 @@ r"""The modality of the model"""
 CreatePromptPromptsFormat = Literal["url", "b64_json", "text", "json_object"]
 r"""Only supported on `image` models."""
 
-CreatePromptQuality = Literal["standard", "hd"]
-r"""Only supported on `image` models."""
-
 CreatePromptResponseFormatPromptsResponseType = Literal["json_object"]
 
 
@@ -726,7 +720,7 @@ class CreatePromptModelParametersTypedDict(TypedDict):
     r"""Only supported on `image` models."""
     dimensions: NotRequired[str]
     r"""Only supported on `image` models."""
-    quality: NotRequired[CreatePromptQuality]
+    quality: NotRequired[str]
     r"""Only supported on `image` models."""
     style: NotRequired[str]
     r"""Only supported on `image` models."""
@@ -788,7 +782,7 @@ class CreatePromptModelParameters(BaseModel):
     dimensions: Optional[str] = None
     r"""Only supported on `image` models."""
 
-    quality: Optional[CreatePromptQuality] = None
+    quality: Optional[str] = None
     r"""Only supported on `image` models."""
 
     style: Optional[str] = None