+++ packages/orq-rc/src/orq_ai_sdk/models/getonepromptop.py
@@ -45,9 +45,6 @@ r"""The modality of the model"""
 GetOnePromptFormat = Literal["url", "b64_json", "text", "json_object"]
 r"""Only supported on `image` models."""
 
-GetOnePromptQuality = Literal["standard", "hd"]
-r"""Only supported on `image` models."""
-
 GetOnePromptResponseFormatPromptsType = Literal["json_object"]
 
 
@@ -148,7 +145,7 @@ class GetOnePromptModelParametersTypedDict(TypedDict):
     r"""Only supported on `image` models."""
     dimensions: NotRequired[str]
     r"""Only supported on `image` models."""
-    quality: NotRequired[GetOnePromptQuality]
+    quality: NotRequired[str]
     r"""Only supported on `image` models."""
     style: NotRequired[str]
     r"""Only supported on `image` models."""
@@ -210,7 +207,7 @@ class GetOnePromptModelParameters(BaseModel):
     dimensions: Optional[str] = None
     r"""Only supported on `image` models."""
 
-    quality: Optional[GetOnePromptQuality] = None
+    quality: Optional[str] = None
     r"""Only supported on `image` models."""
 
     style: Optional[str] = None