+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatapointop.py
@@ -223,5 +223,5 @@ class RetrieveDatapointResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-04-24T06:42:39.159Z")
+    updated: Optional[datetime] = parse_datetime("2025-05-15T06:43:27.397Z")
     r"""The date and time the resource was last updated"""