+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -86,7 +86,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JSK8HMVX4FCB83XK1KDNEKW4"
+        "01JV9AY7KP910J8HKAYQSYR4M5"
     )
     r"""The id of the resource"""